Add memmove to clib

+++ nova/libc/string/memcpy.c
@@ -2,7 +2,7 @@
 #undef memcpy
 
 
-void* memcpy(void *dst, const void* src, int n) {
+void* memcpy(void *dst, const void* src, size_t n) {
     char* dest = (char*)dst;
     const char* source = (char*)src;
